app/requirements.txt
streamlit
streamlit-extras
openai
langchain
SQLAlchemy
snowflake-sqlalchemy
python-dotenv==1.0.0
numpy
pandas
plotly
scikit-learn
pmdarima
matplotlib
pyarrow==10.0.1
tiktoken
chromadb
PyPDF2
faiss-cpu
cx_oracle
google-cloud-aiplatform[preview]
google-auth